style(createcontroller): renaming all to createController
