+++ packages/snap-preact/src/create/createFinderController.ts
