+++ packages/snap-preact/src/create/createSearchController.ts
