+++ packages/snap-preact/src/create/index.ts
@@ -1,4 +1,4 @@
-export { default as createAutocompleteController } from './autocompleteController';
-export { default as createFinderController } from './finderController';
-export { default as createRecommendationsController } from './recommendationController';
-export { default as createSearchController } from './searchController';
+export { default as createAutocompleteController } from './createAutocompleteController';
+export { default as createFinderController } from './createFinderController';
+export { default as createRecommendationsController } from './createRecommendationController';
+export { default as createSearchController } from './createSearchController';
